+++ lib/src/structures/player.dart
@@ -1,0 +1,11 @@
+/// Player object
+class Player {
+  /// account id of the player
+  String accountId;
+
+  /// display name of the player
+  String displayName;
+
+  /// constructor
+  Player(this.accountId, {required this.displayName});
+}